+++ cpp/order/HexaticTranslational.h
@@ -19,7 +19,7 @@
 
 namespace freud { namespace order {
 
-//! Parent class for Hexatic and Translational
+//! Parent class for Hexatic
 /*!
  */
 template<typename T> class HexaticTranslational
@@ -73,23 +73,6 @@ public:
     ~Hexatic() override = default;
 
     //! Compute the hexatic order parameter
-    void compute(const freud::locality::NeighborList* nlist, const freud::locality::NeighborQuery* points,
-                 freud::locality::QueryArgs qargs);
-};
-
-//! Compute the translational order parameter for a set of points
-/*! THIS CLASS IS DEPRECATED AND WILL BE REMOVED IN THE NEXT MAJOR RELEASE OF FREUD.
- */
-class Translational : public HexaticTranslational<float>
-{
-public:
-    //! Constructor
-    Translational(float k = 6, bool weighted = false);
-
-    //! Destructor
-    ~Translational() override = default;
-
-    //! Compute the translational order parameter
     void compute(const freud::locality::NeighborList* nlist, const freud::locality::NeighborQuery* points,
                  freud::locality::QueryArgs qargs);
 };