+++ freud/order.pyx
@@ -412,94 +412,6 @@ cdef class Hexatic(_PairCompute):
             return None
 
 
-cdef class Translational(_PairCompute):
-    r"""Compute the translational order parameter for each particle.
-
-    The translational order parameter is used to measure order in the bonds
-    of 2D systems. The translational order parameter for a particle :math:`i`
-    and its :math:`N_b` neighbors :math:`j` is given by a sum over the
-    neighbors, treating the 2D vectors between each pair of particles as a
-    complex number with real part corresponding to the x-component of the
-    vector and imaginary part corresponding to the y-component of the vector,
-    divided by a normalization constant :math:`k`:
-
-    :math:`\psi\left( i \right) = \frac{1}{k} \sum \limits_{j=1}^{N_b} x_{ij} + y_{ij} i`
-
-    The translational order parameter as written above is **complex-valued**.
-
-    .. note::
-        **2D:** :class:`freud.order.Translational` is only defined for 2D
-        systems. The points must be passed in as :code:`[x, y, 0]`.
-
-    .. note::
-        This class is slated for deprecation and will be removed in freud 3.0.
-
-    Args:
-        k (float, optional):
-            Normalization of order parameter (Default value = :code:`6.0`).
-    """  # noqa E501
-    cdef freud._order.Translational * thisptr
-
-    def __cinit__(self, k=6.0):
-        warnings.warn("This class is deprecated and will be removed in "
-                      "version 3.0", FreudDeprecationWarning)
-        self.thisptr = new freud._order.Translational(k, False)
-
-    def __dealloc__(self):
-        del self.thisptr
-
-    def compute(self, system, neighbors=None):
-        r"""Calculates the local descriptors.
-
-        Args:
-            system:
-                Any object that is a valid argument to
-                :class:`freud.locality.NeighborQuery.from_system`.
-            neighbors (:class:`freud.locality.NeighborList` or dict, optional):
-                Either a :class:`NeighborList <freud.locality.NeighborList>` of
-                neighbor pairs to use in the calculation, or a dictionary of
-                `query arguments
-                <https://freud.readthedocs.io/en/stable/topics/querying.html>`_
-                (Default value: None).
-        """
-        cdef:
-            freud.locality.NeighborQuery nq
-            freud.locality.NeighborList nlist
-            freud.locality._QueryArgs qargs
-            const float[:, ::1] l_query_points
-            unsigned int num_query_points
-
-        nq, nlist, qargs, l_query_points, num_query_points = \
-            self._preprocess_arguments(system, neighbors=neighbors)
-
-        self.thisptr.compute(nlist.get_ptr(),
-                             nq.get_ptr(), dereference(qargs.thisptr))
-        return self
-
-    @property
-    def default_query_args(self):
-        """The default query arguments are
-        :code:`{'mode': 'nearest', 'num_neighbors': int(self.k)}`."""
-        return dict(mode="nearest", num_neighbors=int(self.k))
-
-    @_Compute._computed_property
-    def particle_order(self):
-        """:math:`\\left(N_{particles} \\right)` :class:`numpy.ndarray`: Order
-        parameter."""
-        return freud.util.make_managed_numpy_array(
-            &self.thisptr.getOrder(),
-            freud.util.arr_type_t.COMPLEX_FLOAT)
-
-    @property
-    def k(self):
-        """float: Normalization of the order parameter."""
-        return self.thisptr.getK()
-
-    def __repr__(self):
-        return "freud.order.{cls}(k={k})".format(
-            cls=type(self).__name__, k=self.k)
-
-
 cdef class Steinhardt(_PairCompute):
     r"""Compute one or more of the rotationally invariant Steinhardt order
     parameter :math:`q_l` or :math:`w_l` for a set of points
@@ -1071,3 +983,154 @@ cdef class RotationalAutocorrelation(_Compute):
     def __repr__(self):
         return "freud.order.{cls}(l={sph_l})".format(cls=type(self).__name__,
                                                      sph_l=self.l)
+
+
+cdef class ContinuousCoordination(_PairCompute):
+    r"""Computes the continuous local coordination number.
+
+    The :class:`ContinuousCoordination` class implements extensions of the Voronoi
+    discrete coordination number to the real numbers. The formulas for the
+    various implementations are:
+
+    Power:
+
+    .. math::
+
+        CN_p = N^{2.0 - m} \sum_{i=1}^{k}
+        {\left[\left(\frac{V_i}{V}\right)^{m}\right]}^{-1}
+
+    Log:
+
+    .. math::
+
+        CN_{log} = \frac{-1}{\log{N}} \sum_{i=1}^{k}\log{\left(\frac{V_i}{V}\right)}
+
+    Exponential:
+
+    .. math::
+
+        CN_{exp} = \sum_{i=1}^{k}\exp{\left(\frac{V_i}{V} - \frac{1}{N} \right)}
+
+    where :math:`k` is the number of neighbors a particle has, :math:`V_i` is
+    the volume of the pyramid (or area of the triangle in 2D) whose base is the
+    Voronoi polytope facet between the central particle and neighbor :math:`i`
+    and whose height is half the distance vector, and :math:`V` is the
+    volume/area of the Voronoi polytope.
+
+    Note:
+        When using multiple powers, space them out to avoid high correlations. A
+        minimum spacing of 2.0 is recommended with even larger values leading to
+        less correlation.
+
+    Args:
+        powers (list[float], optional): The powers to compute the continuous
+            coordination number for. The default value indicates only compute
+            for power 2.0.
+            (Default value: None)
+        compute_log (`bool`, optional): Whether to compute the log continuous
+            coordination number.
+            (Default value: :code:`True`)
+        compute_exp (`bool`, optional): Whether to compute the exp continuous
+            coordination number.
+            (Default value: :code:`True`)
+    """
+    cdef freud._order.ContinuousCoordination* thisptr
+
+    def __cinit__(self, powers=None, compute_log=True, compute_exp=True):
+        if powers is None:
+            powers = [2.0]
+        self.thisptr = new freud._order.ContinuousCoordination(
+            powers, compute_log, compute_exp)
+
+    def __dealloc__(self):
+        del self.thisptr
+
+    def compute(self, system=None, voronoi=None):
+        r"""Calculates the local coordination number for the specified points.
+
+        Example::
+
+            >>> import freud
+            >>> box, points = freud.data.make_random_system(10, 100, seed=0)
+            >>> # Compute ContinuousCoordination
+            >>> coord = freud.order.ContinuousCoordination([2, 4], True)
+            >>> coord.compute(system=(box, points))
+            freud.order.ContinuousCoordination(...)
+
+        Args:
+            system (optional):
+                Any object that is a valid argument to
+                :class:`freud.locality.NeighborQuery.from_system`.
+                (Default value: None).
+            voronoi (:class:`freud.locality.Voronoi`, optional):
+                A precomputed Voronoi compute object. If provided, the object is
+                assumed to have been computed already, and system is ignored.
+                (Default value: None).
+        """
+        cdef freud.locality.Voronoi cpp_voronoi
+        if system is None and voronoi is None:
+            raise ValueError("Must specify system or voronoi.")
+        if voronoi is None:
+            voronoi = freud.locality.Voronoi()
+            voronoi.compute(system)
+        elif not hasattr(voronoi, "nlist"):
+            raise RuntimeError(
+                "Must call compute on Voronoi object prior to computing coordination.")
+        cpp_voronoi = voronoi
+        self.thisptr.compute(cpp_voronoi.thisptr)
+        return self
+
+    @_Compute._computed_property
+    def particle_order(self):
+        """(:math:`(N_{points}, N_{coord}`) :class:`numpy.ndarray`: \
+                coordination of points per query point.
+
+        Coordination numbers are in order of selected powers, log, and exp.
+        """
+        return self.coordination
+
+    @_Compute._computed_property
+    def coordination(self):
+        """(:math:`(N_{points}, N_{coord}`) :class:`numpy.ndarray`): \
+                coordination of points per query point.
+
+        Coordination numbers are in order of selected powers, log, and exp.
+        """
+        return freud.util.make_managed_numpy_array(
+            &self.thisptr.getCoordination(),
+            freud.util.arr_type_t.FLOAT)
+
+    @property
+    def powers(self):
+        """list[float]: The powers to compute the continuous coordination number.
+
+        Changes to this property are not reflected when computing coordination
+        numbers.
+        """
+        return self.thisptr.getPowers()
+
+    @property
+    def compute_log(self):
+        """bool: Whether to compute the log continuous coordination number."""
+        return self.thisptr.getComputeLog()
+
+    @property
+    def compute_exp(self):
+        """bool: Whether to compute the exponential coordination number."""
+        return self.thisptr.getComputeExp()
+
+    @property
+    def number_of_coordinations(self):
+        """int: The number of coordination numbers computed."""
+        return self.thisptr.getNumberOfCoordinations()
+
+    def __repr__(self):
+        return (
+            "freud.order.{cls}(powers={powers}, "
+            "compute_log={compute_log}, compute_exp={compute_exp})"
+        ).format(
+            cls=type(self).__name__,
+            powers=self.powers,
+            compute_log=self.compute_log,
+            compute_exp=self.compute_exp,
+        )